Add sources to the topic explorer
